+++ src/app/controllers/SiteController.js
@@ -1,0 +1,8 @@
+class SiteController {
+  // get interface
+  interface(req, res) {
+    res.render("homepage");
+  }
+}
+
+module.exports = new SiteController();
