--- src/app/controllers/index.js
@@ -1,1 +1,0 @@
-console.log("haha");
