+++ src/routers/index.js
@@ -1,0 +1,7 @@
+const siteRouter = require("./siteRouter");
+
+function routers(app) {
+  app.use("/", siteRouter);
+}
+
+module.exports = routers;
